Add files via upload
This class will store a username and a password for an instance of the class. The "this" keyword is used so that the function "save_user" uses this instance of the $username and $password properties. The output will be a saved username and password once the results are "echoed" unto the webpage.