Add files via upload
This program demonstrates that using the clone method can enable us to clone an object and enable the second object to be its own individual object. If we did not use the clone method, the two objects would reference the same object and changing the properties of one object would change the properties of another. This is shown in PHP.